Split off class Writ to own file
